Refactor socket listener out of control socket. The control socket can be seperated into the code to listen to the control socket, and the code to handle the requests recived on it (i.e. exposing controller metrics).
The socket listener can be used elsewhere so it is moved to the internal
package.
