+++ internal/worker/controlsocket/manifold.go
@@ -1,4 +1,4 @@
-// Copyright 2023 Canonical Ltd.
+// Copyright 2024 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
 package controlsocket
@@ -10,17 +10,28 @@ import (
 	"github.com/juju/worker/v4"
 	"github.com/juju/worker/v4/dependency"
 
+	"github.com/juju/juju/internal/socketlistener"
 	"github.com/juju/juju/internal/worker/common"
 	workerstate "github.com/juju/juju/internal/worker/state"
 	"github.com/juju/juju/state"
 )
 
+// SocketListener describes a worker that listens on a unix socket.
+type SocketListener interface {
+	worker.Worker
+}
+
+func NewSocketListener(config socketlistener.Config) (SocketListener, error) {
+	return socketlistener.NewSocketListener(config)
+}
+
 // ManifoldConfig describes the dependencies required by the controlsocket worker.
 type ManifoldConfig struct {
-	StateName  string
-	Logger     Logger
-	NewWorker  func(Config) (worker.Worker, error)
-	SocketName string
+	StateName         string
+	Logger            Logger
+	NewWorker         func(Config) (worker.Worker, error)
+	NewSocketListener func(socketlistener.Config) (SocketListener, error)
+	SocketName        string
 }
 
 // Manifold returns a Manifold that encapsulates the controlsocket worker.
@@ -43,6 +54,9 @@ func (cfg ManifoldConfig) Validate() error {
 	}
 	if cfg.NewWorker == nil {
 		return errors.NotValidf("nil NewWorker func")
+	}
+	if cfg.NewSocketListener == nil {
+		return errors.NotValidf("nil NewSocketListener func")
 	}
 	if cfg.SocketName == "" {
 		return errors.NotValidf("empty SocketName")
@@ -75,9 +89,10 @@ func (cfg ManifoldConfig) start(ctx context.Context, getter dependency.Getter) (
 
 	var w worker.Worker
 	w, err = cfg.NewWorker(Config{
-		State:      stateShim{st},
-		Logger:     cfg.Logger,
-		SocketName: cfg.SocketName,
+		State:             stateShim{st},
+		Logger:            cfg.Logger,
+		SocketName:        cfg.SocketName,
+		NewSocketListener: cfg.NewSocketListener,
 	})
 	if err != nil {
 		return nil, errors.Trace(err)
